editor: move completion functions
