+++ enter/functions.h
@@ -24,7 +24,9 @@
 #define MUTT_ENTER_FUNCTIONS_H
 
 #include <stdbool.h>
+#include <stddef.h>
 
+struct EnterState;
 struct EnterWindowData;
 struct MuttWindow;
 
@@ -50,5 +52,6 @@ struct EnterFunction
 
 int enter_function_dispatcher(struct MuttWindow *win, int op);
 bool self_insert(struct EnterWindowData *wdata, int ch);
+void replace_part(struct EnterState *es, size_t from, const char *buf);
 
 #endif /* MUTT_ENTER_FUNCTIONS_H */
